chore(script-fixes): script updates and bootstrap-tools modularization
