+++ scripts/utils/vault.ts
@@ -1,0 +1,26 @@
+import hre from "hardhat";
+import { SignerWithAddress } from "@nomiclabs/hardhat-ethers/dist/src/signer-with-address";
+
+import { VaultFactory, AssetVault } from "../../typechain";
+
+type Signer = SignerWithAddress;
+
+ let vault: AssetVault | undefined;
+export const createVault = async (factory: VaultFactory, user: Signer): Promise<AssetVault> =>     {
+    const tx = await factory.connect(user).initializeBundle(await user.getAddress());
+    const receipt = await tx.wait();
+
+    if (receipt && receipt.events) {
+        for (const event of receipt.events) {
+            if (event.args && event.args.vault) {
+                vault = <AssetVault>await hre.ethers.getContractAt("AssetVault", event.args.vault);
+            }
+        }
+    } else {
+        throw new Error("Unable to create new vault");
+    }
+    if (!vault) {
+        throw new Error("Unable to create new vault");
+    }
+    return vault;
+    };
